Add commented-out transaction attribute with TODO note
Added the `[HandlerDatabaseTransactionAttribute<CoreDbContext>]`
attribute to the `CustomerCreateCommandHandler` class, but left it
commented out with a `TODO` note. The note explains that the attribute
cannot currently be used due to `CoreDbContext` being defined in the
`Infrastructure` layer, which creates a dependency issue. This change
indicates a future intention to implement database transaction handling
once the dependency problem is resolved.